requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0